+++ NEWS.md
@@ -1,5 +1,10 @@
 # dplyr (development version)
 
+* `recode()` is superseded in favor of `case_match()`. `recode_factor()` is
+  superseded as well, but we don't have a direct replacement for it yet. We plan
+  to add one to forcats, but in the meantime you can often use a pattern of
+  `case_match(.ptype = factor(levels = ))` instead (#6433).
+
 * `across()` has gained a new experimental `.unpack` argument to optionally
   unpack (as in, `tidyr::unpack()`) data frames returned by functions in `.fns`
   (#6360).